requirements.txt
streamlit
torch
torchvision
Pillow
python-docx
git+https://github.com/lukas-blecher/LaTeX-OCR.git
